# Conflicts:
#	backend/pom.xml
#	backend/src/main/java/ShapeSlate/backend/controllers/UserController.java
#	backend/src/main/java/ShapeSlate/backend/models/SlateUser.java
#	backend/src/main/java/ShapeSlate/backend/services/UserService.java
#	backend/src/main/resources/application.properties
#	frontend/src/app/_helpers/login-auth.guard.ts
#	frontend/src/app/_models/slateuser.ts
#	frontend/src/app/_services/account.service.ts
#	frontend/src/app/account/layout.component.ts
#	frontend/src/app/app-routing.module.ts
#	frontend/src/app/app.component.html
#	frontend/src/app/app.component.ts
#	frontend/src/app/board/board.component.css
#	frontend/src/app/message-box/message-box.component.ts

+++ backend/src/main/java/ShapeSlate/backend/controllers/BoardController.java
@@ -1,9 +1,8 @@
 package ShapeSlate.backend.controllers;
 
-import ShapeSlate.backend.models.Board;
-import ShapeSlate.backend.models.CanvasUpdateComparator;
-import ShapeSlate.backend.models.CanvasWhiteboardShapeOptions;
-import ShapeSlate.backend.models.CanvasWhiteboardUpdate;
+import ShapeSlate.backend.models.*;
+import ShapeSlate.backend.models.comparators.CanvasUpdateComparatorById;
+import ShapeSlate.backend.models.comparators.CanvasUpdateComparatorByType;
 import ShapeSlate.backend.services.BoardService;
 import ShapeSlate.backend.services.CanvasWhiteboardShapeOptionsService;
 import ShapeSlate.backend.services.CanvasWhiteboardUpdateService;
@@ -11,15 +10,10 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @CrossOrigin(origins = "http://localhost:4200")
 @RestController
@@ -33,12 +27,12 @@ public class BoardController {
 
     @PostMapping("/board")
     public List<Board> save(@RequestBody String json) throws JsonProcessingException {
-//        System.out.println("save request!");
         ObjectMapper mapper = new ObjectMapper();
         mapper.configure(DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY, true);
         List<Board> myBoard = Arrays.asList(mapper.readValue(json, Board[].class));
 
         for (Board aBoard : myBoard) {
+            boardService.save(aBoard);
             List<CanvasWhiteboardUpdate> theUpdates = aBoard.getCanvasWhiteboardUpdates();
             if (theUpdates != null) {
                 for (CanvasWhiteboardUpdate anUpdate : theUpdates) {
@@ -50,7 +44,6 @@ public class BoardController {
             }
             canvasWhiteboardUpdateService.saveAll(theUpdates);
         }
-//        System.out.println(myBoard.get(0).getCanvasWhiteboardUpdates());
         return (List<Board>) boardService.saveAll(myBoard);
     }
 
@@ -68,12 +61,14 @@ public class BoardController {
         // get all updates
         List<CanvasWhiteboardUpdate> updates = (List<CanvasWhiteboardUpdate>) canvasWhiteboardUpdateService.findByBoardId(myBoard);
         // and delete them
-        for (CanvasWhiteboardUpdate anUpdate : updates) {
-            CanvasWhiteboardShapeOptions shapeOptions = anUpdate.getSelectedShapeOptions();
-            if (shapeOptions != null) {
-            canvasWhiteboardShapeOptionsService.delete(anUpdate.getSelectedShapeOptions());
+        if (updates != null) {
+            for (CanvasWhiteboardUpdate anUpdate : updates) {
+                CanvasWhiteboardShapeOptions shapeOptions = anUpdate.getSelectedShapeOptions();
+                canvasWhiteboardUpdateService.delete(anUpdate);
+                if (shapeOptions != null) {
+                    canvasWhiteboardShapeOptionsService.delete(anUpdate.getSelectedShapeOptions());
+                }
             }
-            canvasWhiteboardUpdateService.delete(anUpdate);
         }
     }
 
@@ -84,16 +79,79 @@ public class BoardController {
 
     @GetMapping("/board/{id}")
     public Board BoardById(@PathVariable int id) {
+        // BUG (probably); early termination of updateset when multiple people are drawing.
+        // Too sad too bad that ng2-canvas-whiteboard NEEDS type 2 at the end of a changeset...
+
         Board myBoard = boardService.findById(id).orElse(null);
         if (myBoard != null) {
             List<CanvasWhiteboardUpdate> myUpdates = myBoard.getCanvasWhiteboardUpdates();
             if (myUpdates != null) {
-                myUpdates.sort(new CanvasUpdateComparator());
-                myBoard.setCanvasWhiteboardUpdates(myUpdates);
-//                System.out.println(myUpdates.get(myUpdates.size()-1).getType());
+                List<CanvasWhiteboardUpdate> mySortedUpdates = new ArrayList<>();
+
+                // Sort by ID so that updates are sort of in the right order
+                myUpdates.sort(new CanvasUpdateComparatorById());
+
+                // Get unique UUIDs
+                String lastUUID = "";
+                Set<String> uniqueUUIDs = new HashSet<>();
+                for (CanvasWhiteboardUpdate update : myUpdates) {
+                    uniqueUUIDs.add(update.getUUID());
+                    lastUUID = update.getUUID();
+                }
+                // grab set of updates for each UUID
+                for (String UUID: uniqueUUIDs) {
+                    List<CanvasWhiteboardUpdate> changeSet = new ArrayList<>();
+                    for (CanvasWhiteboardUpdate update: myUpdates) {
+                        if(update.getUUID().equals(UUID)) {
+                            changeSet.add(update);
+                        }
+                    }
+                    // sort the changeset by type
+                    changeSet.sort(new CanvasUpdateComparatorByType());
+                    List<CanvasWhiteboardUpdate> changeSetBegin = new ArrayList<>();
+                    List<CanvasWhiteboardUpdate> changeSetInBetween = new ArrayList<>();
+                    List<CanvasWhiteboardUpdate> changeSetEnd = new ArrayList<>();
+                    for (CanvasWhiteboardUpdate update: changeSet) {
+                        if(update.getType() == 0) {
+                            changeSetBegin.add(update);
+                        } else if (update.getType() == 1) {
+                            changeSetInBetween.add(update);
+                        } else if (update.getType() == 2) {
+                            changeSetEnd.add(update);
+                        }
+                    }
+                    changeSetBegin.sort(new CanvasUpdateComparatorById());
+                    changeSetInBetween.sort(new CanvasUpdateComparatorById());
+                    changeSetEnd.sort(new CanvasUpdateComparatorById());
+
+                    if (!changeSetBegin.isEmpty()) {
+                        boolean shouldAdd = true;
+
+                        if (changeSetEnd.isEmpty()) {
+                            CanvasWhiteboardUpdate endUpdate;
+                            if (changeSetInBetween.isEmpty()) {
+                                endUpdate = changeSetBegin.get(changeSetBegin.size()-1);
+                            } else {
+                                endUpdate = changeSetInBetween.get(changeSetInBetween.size()-1);
+                            }
+
+                            if (!endUpdate.getUUID().equals(lastUUID)) {
+                                endUpdate.setType(2);
+                                changeSetEnd.add(endUpdate);
+                            } else {shouldAdd = false;}
+                        }
+
+                        if (shouldAdd) {
+                            mySortedUpdates.addAll(changeSetBegin);
+                            mySortedUpdates.addAll(changeSetInBetween);
+                            mySortedUpdates.addAll(changeSetEnd);
+                        }
+
+                    }
+                }
+                myBoard.setCanvasWhiteboardUpdates(mySortedUpdates);
             }
         }
-
 
         return myBoard;
     }
